(working on bug 113)
Fixed a formerly unnoticed crash bug. If user selected a subset of the dependency files list, then clicked "Copy Selected Audio Into Project", it removed the selected dependency/dependencies, but then crashed on bad reference trying to delete the non-selected dependency/dependencies. This was due to a consistent misuse of wxHashMap::operator[] to check for the existence of a hash element. If you call it with a key that's not in the hash map, yes it returns NULL, the first time, but it creates an entry for that key. So if you test it again with the same key it returns a bad entry. Replaced it with a correct call to wxHashMap::count(). Leland had fixed one of these, but there were 4 more. 

RemoveDependencies() always returned true, and was never checked, so made it return nothing.

More readability improvements.