Fix Bug 504 (P2) - PPC Macs crash importing MP3 files with "guess_file_type" error
in id3.c: psf->fileoffset is a sf_count_t, which is int64_t, which is
larger than size_t on a G4 PPC.
Since psf_binheader_readf expects that argument to be the smaller
size_t, it ends up seeking over and over to the zero position since
the larger bits of the variable are all zeros.
Patch simply casts to size_t.

Patch sent upstream as well.